+++ TomasAI.IFM.Domain.MarketData.Feed.Shared/ViewModels/FuturesTickHLVDataReadModel.cs
@@ -5,4 +5,4 @@ public record FuturesTickHLVDataReadModel(
     DateOnly ValueDate,
     decimal HighPrice,
     decimal LowPrice,
-    int  Volume);
+    long Volume);
