+++ TomasAI.IFM.Domain.MarketData.Shared/ViewModels/FuturesEodDataReadModel.cs
@@ -48,7 +48,7 @@ public record FuturesEodDataV2ReadModel
 
     /// <summary>Session volume.</summary>
     [Key(7)]
-    public int Volume { get; init; }
+    public long Volume { get; init; }
 
     /// <summary>Percent change over the analysis window (daily or rolling).</summary>
     [Key(8)]
@@ -132,7 +132,7 @@ public record FuturesEodDataV2ReadModel
         decimal highPrice,
         decimal lowPrice,
         decimal closePrice,
-        int volume,
+        long volume,
         double dailyPercentChange = 0,
         double dailyStdDev = 0,
         double dailyStdDevAmount = 0,
